+++ lib/src/code/code_generator_class.dart
@@ -1,0 +1,12 @@
+import 'package:fgen/src/code/generate_code.dart';
+
+abstract class CodeGeneratorClass {
+  String get folder;
+  String get fileSuffix;
+
+  String generate({
+    required String name,
+    String? projectName,
+    Set<GenerateType>? allGeneratedTypes,
+  });
+}