+++ test/code/code_helper.dart
@@ -1,11 +1,14 @@
 import 'dart:io';
 
-import 'package:fgen/code/generate_code.dart';
-import 'package:fgen/util/string_extension.dart';
+import 'package:fgen/src/code/generate_code.dart';
+import 'package:fgen/src/util/string_extension.dart';
 import 'package:test/test.dart';
 
 class CodeGeneratorTestHelper {
-  static void testCodeGenerator(GenerateType type, String name, String resultFileName, {bool overwriteOutputFiles = false, List<GenerateType> additionalTypes = const []}) {
+  static void testCodeGenerator(
+      GenerateType type, String name, String resultFileName,
+      {bool overwriteOutputFiles = false,
+      List<GenerateType> additionalTypes = const []}) {
     final file = File('test/code/outputs/$resultFileName.txt');
     final actual = GenerateCode.getGenerator(type).generate(
       name: name.toCamelCase(),