+++ test/util/string_extension_test.dart
@@ -1,4 +1,4 @@
-import 'package:fgen/util/string_extension.dart';
+import 'package:fgen/src/util/string_extension.dart';
 import 'package:test/test.dart';
 
 void main() {