+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.2 - 2023/01/07
+
+- Updated to use MainNavigator as singleton
+- Updated to use get_x_navigation_generator_interface for navigation generation
+
 ## 0.0.1+4 - 19/04/22
 
 - Run as executable